add `<leader>lR` mapping for LSP references
While I did not use the LSP type definition, definition, and implementation functions yet, the references function, which supposedly searches through the code base to find lines that uses said function or class is not actually triggered by `gr`, but instead `lR`. Maybe the documentation is out of sync with the code changes?